# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "My Golang API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/rentals": {
            "get": {
                "description": "Получение списка всех команд",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Аренда"
                ],
                "summary": "Возвращает список всех команд",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RentalView"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Bad"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "Internal"
                        }
                    }
                }
            },
            "post": {
                "description": "Создание новой команды",
                "tags": [
                    "Аренда"
                ],
                "summary": "Создание новой команды",
                "parameters": [
                    {
                        "description": "Данные для создания новой команды",
                        "name": "createRental",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateRentalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.RentalView"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "Unprocessable"
                        }
                    }
                }
            }
        },
        "/api/rentals/{id}": {
            "get": {
                "description": "Получение информации о команде по идентификатору",
                "tags": [
                    "Аренда"
                ],
                "summary": "Возвращает информацию о команде по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID команды",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RentalView"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Bad"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "Not"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаление команды по идентификатору",
                "tags": [
                    "Аренда"
                ],
                "summary": "Удаляет команду по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID команды",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "No"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "Not"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateRentalRequest": {
            "type": "object",
            "required": [
                "end_date",
                "field_id",
                "start_date",
                "team_id"
            ],
            "properties": {
                "comment": {
                    "description": "Статус аренды",
                    "type": "string"
                },
                "end_date": {
                    "description": "Дата завершения аренды",
                    "type": "string"
                },
                "field_id": {
                    "description": "Идентификатор площадки",
                    "type": "integer"
                },
                "id": {
                    "description": "Идентификатор",
                    "type": "integer"
                },
                "start_date": {
                    "description": "Дата начала аренды",
                    "type": "string"
                },
                "team_id": {
                    "description": "Идентификатор команды",
                    "type": "integer"
                }
            }
        },
...
        "models.FieldView": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display": {
                    "type": "boolean"
                },
                "dressing": {
                    "type": "boolean"
                },
                "for_disabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "info": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "logo": {
                    "description": "Логотип",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Media"
                        }
                    ]
                },
                "media": {
                    "description": "Медиа",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Media"
                    }
                },
                "name": {
                    "type": "string"
                },
                "parking": {
                    "type": "boolean"
                },
                "places": {
                    "type": "integer"
                },
                "square": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "toilet": {
                    "type": "boolean"
                }
            }
        },
        "models.Media": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "ext": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "models.RentalView": {
            "type": "object",
            "properties": {
                "comment": {
                    "description": "Комментарий",
                    "type": "string"
                },
                "created_at": {
                    "description": "Дата создания",
                    "type": "string"
                },
                "duration": {
                    "description": "Длительность аренды (например, в часах)",
                    "type": "integer"
                },
                "end_date": {
                    "description": "Дата завершения аренды",
                    "type": "string"
                },
                "field": {
                    "description": "Пощадка",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.FieldView"
                        }
                    ]
                },
                "id": {
                    "description": "Идентификатор",
                    "type": "integer"
                },
                "start_date": {
                    "description": "Дата начала аренды",
                    "type": "string"
                },
                "status": {
                    "description": "Статус аренды",
                    "type": "integer"
                },
                "team": {
                    "description": "Команда",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TeamView"
                        }
                    ]
                },
                "user": {
                    "description": "Пользователь",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.UserView"
                        }
                    ]
                }
            }
        },
        "models.TeamView": {
            "type": "object",
            "properties": {
                "city": {
                    "description": "Город",
                    "type": "string"
                },
                "created_at": {
                    "description": "Дата создания",
                    "type": "string"
                },
                "description": {
                    "description": "Описание",
                    "type": "string"
                },
                "disability_category": {
                    "description": "Категория инвалидности",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "description": "Логотип",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Media"
                        }
                    ]
                },
                "media": {
                    "description": "Медиа",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Media"
                    }
                },
                "name": {
                    "description": "Название",
                    "type": "string"
                },
                "participant_count": {
                    "description": "Кол-во участников",
                    "type": "integer"
                },
                "responsible": {
                    "description": "Ответственный",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.UserView"
                        }
                    ]
                },
                "status": {
                    "description": "Статус",
                    "type": "integer"
                },
                "uniform_color": {
                    "description": "Цвет формы",
                    "type": "string"
                }
            }
        },
...
        "models.UserView": {
            "type": "object",
            "properties": {
                "city": {
                    "description": "Город",
                    "type": "string"
                },
                "created_at": {
                    "description": "Дата создания",
                    "type": "string"
                },
                "email": {
                    "description": "Email",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "description": "Логотип",
                    "type": "string"
                },
                "media": {
                    "description": "Медиа",
                    "type": "string"
                },
                "name": {
                    "description": "ФИО",
                    "type": "string"
                },
                "phone": {
                    "description": "Телефон",
                    "type": "string"
                },
                "status": {
                    "description": "Статус",
                    "type": "integer"
                }
            }
        },
...
    }
}